+++ api/controllers/requests/queue.js
@@ -16,6 +16,9 @@ module.exports = {
     sails.log.debug('Controller requests/queue called.')
 
     try {
+      // Prevent queuing requests if host is lockToDJ and the specified lockToDJ is not on the air
+      if (this.req.payload.lockToDJ !== null && this.req.payload.lockToDJ !== sails.models.meta['A'].dj) { return exits.error(new Error('You are not authorized to queue track requests because you are not on the air.')) }
+
       // Queue the request
       var response = await sails.helpers.requests.queue(1, false, false, inputs.ID)