+++ api/controllers/songs/queue-psa.js
@@ -15,6 +15,9 @@ module.exports = {
   fn: async function (inputs, exits) {
     sails.log.debug('Controller songs/queue-psa called.')
     try {
+      // Prevent adding tracks if host is lockToDJ and the specified lockToDJ is not on the air
+      if (this.req.payload.lockToDJ !== null && this.req.payload.lockToDJ !== sails.models.meta['A'].dj) { return exits.error(new Error('You are not authorized to queue a PSA because you are not on the air.')) }
+
       // Queue applicable PSA
       await sails.helpers.songs.queue(sails.config.custom.subcats.PSAs, 'Top', 1, true, inputs.duration)